+++ src/chestnut/engine/graphics/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(algorithms)
 add_subdirectory(opengl)
 add_subdirectory(renderers)
 
@@ -6,15 +5,8 @@ target_sources(${PROJECT_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/animation2d.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/camera2d.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/camera2d.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/camera3d.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/camera3d.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/colored_polygon2d.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/colored_polygon2d.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/light.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/light.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/material.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/sprite.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/sprite.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/text.hpp
