+++ src/chestnut/engine/graphics/renderers/CMakeLists.txt
@@ -1,10 +1,6 @@
 target_sources(${PROJECT_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/colored_polygon2d_renderer.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/colored_polygon2d_renderer.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/phong_renderer.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/phong_renderer.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/plain_mesh_renderer.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/plain_mesh_renderer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/renderer.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/renderer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/sprite_renderer.hpp
