+++ src/chestnut/engine/resources/CMakeLists.txt
@@ -7,10 +7,6 @@ target_sources(${PROJECT_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/font_resource.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/image_data_resource.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/image_data_resource.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_data_resource.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_data_resource.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_resource.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_resource.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/resource_manager.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/resource_manager.inl
     ${CMAKE_CURRENT_SOURCE_DIR}/resource_manager.cpp
