+++ src/chestnut/engine_extensions/graphics3d/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME}_Graphics3D PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/macros.hpp
+)
+
+add_subdirectory(graphics)
+add_subdirectory(resources)
