+++ src/chestnut/engine_extensions/graphics3d/graphics/algorithms/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(${PROJECT_NAME} PRIVATE
+target_sources(${PROJECT_NAME}_Graphics3D PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/extrude_profile.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/extrude_profile.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/heightmap.hpp
