+++ src/chestnut/engine_extensions/graphics3d/graphics/algorithms/heightmap.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 
-#include "chestnut/engine/macros.hpp"
-#include "chestnut/engine/resources/mesh_data_resource.hpp"
+#include "chestnut/engine_extensions/graphics3d/resources/mesh_data_resource.hpp"
+#include "chestnut/engine_extensions/graphics3d/macros.hpp"
 
 #include <type_traits>
