+++ src/chestnut/engine_extensions/graphics3d/graphics/renderers/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME}_Graphics3D PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/phong_renderer.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/phong_renderer.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/plain_mesh_renderer.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/plain_mesh_renderer.cpp
+)
