+++ src/chestnut/engine_extensions/graphics3d/graphics/renderers/plain_mesh_renderer.cpp
@@ -1,4 +1,4 @@
-#include "chestnut/engine/graphics/renderers/plain_mesh_renderer.hpp"
+#include "chestnut/engine_extensions/graphics3d/graphics/renderers/plain_mesh_renderer.hpp"
 
 #include "chestnut/engine/resources/resource_manager.hpp"
 #include "chestnut/engine/macros.hpp"
