+++ src/chestnut/engine_extensions/graphics3d/resources/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME}_Graphics3D PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_data_resource.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_data_resource.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_resource.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/mesh_resource.cpp
+)
