+++ tests/extrude_profile_test.cpp
@@ -4,11 +4,11 @@
 
 #include "../src/chestnut/engine/init.hpp"
 #include "../src/chestnut/engine/main/window.hpp"
-#include "../src/chestnut/engine/graphics/renderers/plain_mesh_renderer.hpp"
 #include "../src/chestnut/engine/constants.hpp"
 #include "../src/chestnut/engine/macros.hpp"
 #include "../src/chestnut/engine/maths/angles.hpp"
-#include "../src/chestnut/engine/graphics/algorithms/extrude_profile.hpp"
+#include "../src/chestnut/engine_extensions/graphics3d/graphics/renderers/plain_mesh_renderer.hpp"
+#include "../src/chestnut/engine_extensions/graphics3d/graphics/algorithms/extrude_profile.hpp"
 
 
 using namespace chestnut::engine;
