+++ tests/phong_renderer_test.cpp
@@ -4,10 +4,10 @@
 
 #include "../src/chestnut/engine/init.hpp"
 #include "../src/chestnut/engine/main/window.hpp"
-#include "../src/chestnut/engine/graphics/renderers/phong_renderer.hpp"
 #include "../src/chestnut/engine/constants.hpp"
 #include "../src/chestnut/engine/macros.hpp"
 #include "../src/chestnut/engine/maths/angles.hpp"
+#include "../src/chestnut/engine_extensions/graphics3d/graphics/renderers/phong_renderer.hpp"
 
 
 using namespace chestnut::engine;
